+++ grid2op/Observation/_ObsEnv.py
@@ -94,6 +94,22 @@ class _ObsEnv(BaseEnv):
         self.opp_space_state = None
         self.opp_state = None
 
+        # storage
+        self._storage_current_charge_init = None
+        self._storage_previous_charge_init = None
+        self._action_storage_init = None
+        self._amount_storage_init = None
+        self._amount_storage_prev_init = None
+        self._storage_power_init = None
+
+        # storage unit
+        self._storage_current_charge_init = np.zeros(self.n_storage, dtype=dt_float)
+        self._storage_previous_charge_init = np.zeros(self.n_storage, dtype=dt_float)
+        self._action_storage_init = np.zeros(self.n_storage, dtype=dt_float)
+        self._storage_power_init = np.zeros(self.n_storage, dtype=dt_float)
+        self._amount_storage_init = 0.
+        self._amount_storage_prev_init = 0.
+
     def _init_backend(self,
                       init_grid_path,
                       chronics_handler,
@@ -219,6 +235,7 @@ class _ObsEnv(BaseEnv):
                                                                            "load_q": self._load_q}
                                                              })
         self._backend_action_set += new_state_action
+        self._backend_action_set.storage_power.values[:] = 0.
         self.is_init = True
         self.current_obs.reset()
         self.time_stamp = time_stamp
@@ -231,20 +248,21 @@ class _ObsEnv(BaseEnv):
         update the value of the "time dependant" attributes
         """
         self._times_before_line_status_actionable[:] = np.maximum(self._times_before_line_status_actionable - time_step,
-                                                                  0.)
+                                                                  0)
         self._times_before_topology_actionable[:] = np.maximum(self._times_before_topology_actionable - time_step,
-                                                               0.)
+                                                               0)
 
         still_in_maintenance = (self._duration_next_maintenance > time_step) & (self._time_next_maintenance == 0)
-        reconnected = (self._duration_next_maintenance < time_step) & (self._time_next_maintenance == 0)
+        reconnected = (self._duration_next_maintenance <= time_step) & (self._time_next_maintenance == 0)
         first_ts_maintenance = self._time_next_maintenance == time_step
 
         # powerline that are still in maintenance at this time step
         self._time_next_maintenance[still_in_maintenance] = 0
-        self._duration_next_maintenance[still_in_maintenance] -= 1
+        self._duration_next_maintenance[still_in_maintenance] -= time_step
 
         # powerline that will be in maintenance at this time step
         self._time_next_maintenance[first_ts_maintenance] = 0
+        self._duration_next_maintenance[first_ts_maintenance] -= time_step
 
         # powerline that won't be in maintenance at this time step
         self._time_next_maintenance[reconnected] = -1
@@ -276,6 +294,14 @@ class _ObsEnv(BaseEnv):
         self._backend_action_set.all_changed()
         self._backend_action = copy.deepcopy(self._backend_action_set)
         self._oppSpace._set_state(self.opp_space_state, self.opp_state)
+
+        # storage unit
+        self._storage_current_charge[:] = self._storage_current_charge_init
+        self._storage_previous_charge[:] = self._storage_previous_charge_init
+        self._action_storage[:] = self._action_storage_init
+        self._storage_power[:] = self._storage_power_init
+        self._amount_storage = self._amount_storage_init
+        self._amount_storage_prev = self._amount_storage_prev_init
 
     def simulate(self, action):
         """
@@ -352,7 +378,6 @@ class _ObsEnv(BaseEnv):
             A reference to the environment
         """
         real_backend = env.backend
-        self._reward_helper = env._reward_helper
 
         self._load_p, self._load_q, self._load_v = real_backend.loads_info()
         self._prod_p, self._prod_q, self._prod_v = real_backend.generators_info()
@@ -379,11 +404,19 @@ class _ObsEnv(BaseEnv):
         self.target_dispatch_init[:] = env._target_dispatch
         self.actual_dispatch_init[:] = env._actual_dispatch
         self.opp_space_state, self.opp_state = env._oppSpace._get_state()
-        # TODO check redispatching and simulate are working as intended
-        # TODO also update the status of hazards, maintenance etc.
-        # TODO and simulate also when a maintenance is forcasted!
-        # TODO add the opponent budget here (should decrease with the time step :scared:) -> we really need to address
-        # all that before 1.0.0
+
+        # storage units
+        # TODO this is not time independant... i set up the previous charge of the obs env to be
+        # set current charge of the simulated env on purpose
+        self._storage_current_charge_init[:] = env._storage_current_charge
+        self._storage_previous_charge_init[:] = env._storage_previous_charge
+        self._action_storage_init[:] = env._action_storage
+        self._amount_storage_init = env._amount_storage
+        self._amount_storage_prev_init = env._amount_storage_prev
+        self._storage_power_init[:] = env._storage_power
+
+        # time delta
+        self.delta_time_seconds = env.delta_time_seconds
 
     def get_current_line_status(self):
         return self._line_status == 1