+++ CHANGELOG.rst
@@ -47,6 +47,8 @@ Change Log
   powergrid.
 - [FIXED] `obs.connectivity_matrix` now properly takes into account when two objects are disconnected (before
   it was as if there were connected together)
+- [FIXED] some surprising behaviour when using  `obs.simulate` just before or just after a planned
+  maintenance operation.
 - [ADDED] a convenient function to evaluate the impact (especially on topology) of an action on a state
   (`obs + act`)
 - [ADDED] more consistency in the observation attribute names, they are now `gen_p`, `gen_q` and `gen_v`